docker/pyproject.toml
[tool.poetry]
name = "manager"
version = "0.1.0"
description = "BentoML Docker Images Management CLI"
authors = ["BentoML Team <contact@bentoml.ai>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
Jinja2 = "^3.0.1"
Cerberus = "^1.3.4"
glom = "^20.11.0"
python-dotenv = "^0.19.2"
GitPython = "^3.1.18"
plumbum = "^1.7.2"
attrs = "^21.4.0"
click = "^8.0.4"
rich = "^11.2.0"
typing-extensions = ">=3.10,<4"
pyyaml-include = "^1.2.post2"
simple-di = "^0.1.5"
cattrs = "^1.10.0"
fs = "^2.4.15"


[tool.poetry.dev-dependencies]
black = "22.1.0"
isort = "^5.9.1"
pylint = "^2.8.3"
pytest = "^7.0.1"
autoflake = "^1.4"

[tool.poetry.scripts]
manager = "manager.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
